exceptions: move CalldUninitializedError to plugin_helpers
Why:

* comply with wazo-notebook/file-structure.md
